\name{ch}
\alias{ch}
\title{Chen's Chirp}
\description{Chen's chirp.
}
\usage{data(ch)
}
\format{A vector containing 15,000 observations.
}
\source{
  See discussions in the text of ``Practical Time-Frequency Analysis''.
}
\references{
  Carmona, R. A., W. L. Hwang and B Torresani (1998)
  \emph{Practical Time-Frequency Analysis: Gabor and Wavelet Transforms
    with an Implementation in S}, Academic Press, San Diego.
}
\keyword{datasets}
