\name{wpl}
\alias{wpl}
\title{
Plot Dyadic Wavelet Transform.
}
\description{
Plot dyadic wavelet transform(output of \code{\link{mw}}).
}
\usage{
wpl(dwtrans)
}
\arguments{
\item{dwtrans}{
dyadic wavelet transform (output of \code{\link{mw}}).
}}
\value{

}
\details{
}
\references{
}
\seealso{
\code{\link{mw}}, \code{\link{ext}},\code{\link{epl}}.
}
\keyword{ts}
